Local modifications.

+++ OgreMain/src/OgreFileSystem.cpp
@@ -97,7 +97,7 @@ namespace Ogre {
     void FileSystemArchive::findFiles(const String& pattern, bool recursive, 
         bool dirs, StringVector* simpleList, FileInfoList* detailList)
     {
-        long lHandle, res;
+        intptr_t lHandle, res;
         struct _finddata_t tagData;
 
         // pattern can contain a directory name, separate it from mask
